+++ lib/src/universal_ble_pigeon/universal_ble_pigeon_channel.dart
@@ -294,6 +294,7 @@ extension _UniversalBleScanResultExtension on UniversalBleScanResult {
               ?.map((e) => ManufacturerData(e.companyIdentifier, e.data))
               .toList() ??
           [],
+      serviceData: serviceData ?? {},
     );
   }
 }